Added function getExerciseMaxAverage
Which:
''Computes the average session maximum weight lifted for the given exercise and rep number.''

For each unique date in the dataset this function:
    - gets all sets with that date
    - gets all sets matching the parameter exercise
    - gets all sets matching the parameter reps
    - gets the set with the highest weight for that date
    - appends that weight to a list

After all unique dates are run through, the list of max weights is
averaged, and returned.

It's a bit spagetti at the moment, but it's parameterised so it can run
for any exercise and rep number.