Stub test and function for getting competition dates
I want the app to be able to parse the data and pull out dates of
competition based on whether attempts are noted on that date.

I've stubbed a function for this, and a test. I've had to update
the test data to include competition data. It also includes data
before and after the competition so later I'll be able to test if
we can split data on competitions effectively.